refactor: Перенес папку pages внутрь components
