+++ src/components/pages/Landing/BaseText/BaseText.jsx
