+++ src/components/pages/Landing/Tech/Tech.css
