+++ src/components/pages/Landing/Tech/Tech.jsx
