+++ src/components/pages/Landing/Techs/Techs.css
