+++ src/components/pages/SavedMovies/SavedMovies.jsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react';
 
-import MoviesCardList from '../../components/MoviesCardList/MoviesCardList';
-import SearchMovieForm from '../../components/SearchMovieForm/SearchMovieForm';
-import Preloader from '../../components/ui/Preloader/Preloader';
-import Empty from '../../components/Empty/Empty';
+import MoviesCardList from '../../MoviesCardList/MoviesCardList';
+import SearchMovieForm from '../../SearchMovieForm/SearchMovieForm';
+import Preloader from '../../ui/Preloader/Preloader';
+import Empty from '../../Empty/Empty';
 
-import { MOVIE_BASE_URL } from '../../utils/constants';
+import { MOVIE_BASE_URL } from '../../../utils/constants';
 
 import './SavedMovies.css';
