+++ src/routes/routes.js
@@ -3,13 +3,13 @@ import { createBrowserRouter } from 'react-router-dom';
 
 import App from '../App';
 
-const Landing = lazy(() => import('../pages/Landing/Landing'));
-const NotFound = lazy(() => import('../pages/NotFound/NotFound'));
-const Register = lazy(() => import('../pages/Register/Register'));
-const Profile = lazy(() => import('../pages/Profile/Profile'));
-const Login = lazy(() => import('../pages/Login/Login'));
-const Movies = lazy(() => import('../pages/Movies/Movies'));
-const SavedMovies = lazy(() => import('../pages/SavedMovies/SavedMovies'));
+const Landing = lazy(() => import('../components/pages/Landing/Landing'));
+const NotFound = lazy(() => import('../components/pages/NotFound/NotFound'));
+const Register = lazy(() => import('../components/pages/Register/Register'));
+const Profile = lazy(() => import('../components/pages/Profile/Profile'));
+const Login = lazy(() => import('../components/pages/Login/Login'));
+const Movies = lazy(() => import('../components/pages/Movies/Movies'));
+const SavedMovies = lazy(() => import('../components/pages/SavedMovies/SavedMovies'));
 
 export const router = createBrowserRouter([
   {
